+++ lib/aipp/scopes/shoot/runner.rb
@@ -19,10 +19,8 @@ module AIPP
         parse_sections
         if aixm.features.any?
           validate_aixm
-          write_aixm(AIPP.options.output_file || output_file)
-        else
-          warn("no features to write")
         end
+        write_aixm(AIPP.options.output_file || output_file)
         write_config
       end